refactor: 돈과 관련된 원시값 Money로 리팩토링 및 연산 Money 클래스로 이동

+++ src/main/java/lotto/model/LottoConstants.java
@@ -8,5 +8,5 @@ public final class LottoConstants {
     public static final int LOTTO_NUMBER_SIZE = 6;
     public static final int START_NUMBER = 1;
     public static final int END_NUMBER = 45;
-    public static final int LOTTO_PRICE_UNIT = 1000;
+    public static final long LOTTO_PRICE_UNIT = 1000L;
 }
